Replace withFocusTrap HOC with component with render prop pattern
