+++ src/components/NodeToolbar/AccessibilityEditor/ToiletStatusEditor.js
@@ -6,6 +6,7 @@ import { t } from 'ttag';
 import useImperialUnits from '../../../lib/useImperialUnits';
 import { accessibleToiletDescription } from '../../../lib/Feature';
 import type { WheelmapFeature, YesNoUnknown } from '../../../lib/Feature';
+import FocusTrap from '../../../lib/FocusTrap';
 
 import { saveToiletStatus } from './saveStatus';
 import RadioStatusEditor from './RadioStatusEditor';
@@ -54,7 +55,8 @@ export default function ToiletStatusEditor(props: Props) {
   };
 
   return (
-    <RadioStatusEditor
+    <FocusTrap
+      component={RadioStatusEditor}
       {...props}
       undefinedStringValue="unknown"
       getValueFromFeature={feature => feature.properties.wheelchair_toilet}
@@ -65,6 +67,6 @@ export default function ToiletStatusEditor(props: Props) {
       descriptionForValue={value => descriptions[value]}
     >
       <header id="wheelchair-accessibility-header">{headerText}</header>
-    </RadioStatusEditor>
+    </FocusTrap>
   );
 }
